+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `search.sort_by` can now be set to `false` to disable sorting of search results.
 - Support `workspace/didRenameFiles` and file changes will be prompted with confirmation to rename references.
   - Use `opts.link.auto_update` to control whether ask for confirmation or auto run, false by default.
+- `make docgen` to auto update module docs with up to date module options.
 
 ### Fixed